Part_2
This is the second piece of data that I wanted to show with the first data frame, I ran out of time before I could figure out how to show multiple sets of data frames on the same pandas graph but I thought it might be interesting to see. This set displays placement of pubg players relative to their damage delt in game, looking at the graph of data set one and two, we can see a correlation between the sets of information.